+++ test/escaping_test.js
@@ -23,6 +23,10 @@ describe('escaping', () => {
     it('must escape regular expressions', () => {
       expect(lucene.regex.escape('foo/bar')).to.equal('foo\\/bar');
     });
+
+    it('must escape line breaks', () => {
+      expect(lucene.term.escape('a\nb')).to.equal('a\\\nb');
+    });
   });
 
   describe('unescape', () => {
@@ -39,8 +43,13 @@ describe('escaping', () => {
     it('must unescape phrases', () => {
       expect(lucene.phrase.unescape('foo\\"bar')).to.equal('foo"bar');
     });
+
     it('must unescape regular expressions', () => {
       expect(lucene.regex.unescape('foo\\/bar')).to.equal('foo/bar');
+    });
+
+    it('must unescape line breaks', () => {
+      expect(lucene.term.unescape('a\\\nb')).to.equal('a\nb');
     });
   });
 });